# swagger.yaml
openapi: 3.0.0
info:
  title: Papyro API
  description: Hypermedia project for a library (2019)
  version: 1.0.0
servers:
  - url: https://papyro-hypermedia.herokuapp.com/api
    description: Base URL of the API server
paths:
...
  /events:
    get:
      summary: Returns all events
      tags:
        - events
      parameters:
        - in: query
          name: startDate
          description: only select events from this date (format mm/dd/yyyy)
          schema:
            type: string
            format: date
            example: 6/1/2019
        - in: query
          name: endDate
          description: only select events up to this date (format mm/dd/yyyy)
          schema:
            type: string
            format: date
            example: 7/1/2019
      responses:
        '200':
          description: an array containing all events
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Event'
        '400':
          description: Bad formatted date
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
        '404':
          description: No events in this date range
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
  /events/{id}:
    get:
      summary: Returns the event with the given id
      tags:
        - events
      parameters:
        - in: path
          name: id
          description: The id of the event to get
          schema:
            type: integer
            format: int32
            example: 1
          required: true
      responses:
        '200':
          description: The event details
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Event'
        '400':
          description: Event id must be an integer
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...